# docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /schools/public:
    get:
      description: Retrieves a list of public schools with optional filtering and
        pagination
      parameters:
      - description: Filter by school name
        in: query
        name: name
        type: string
      - description: Filter by city
        in: query
        name: city
        type: string
      - description: Filter by country code
        in: query
        name: country_code
        type: string
      - default: 1
        description: Page number for pagination
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of schools with pagination metadata
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get public schools
      tags:
      - schools
...
swagger: "2.0"